CAPSTONE PROJECT 2025
These are the files included in the Project Handoff for my Capstone project. I was the frontend developer, Sahil Patel, a classmate, was the backend developer, and Feifei Xie did more support work for the frontend and backend. Please unzip and follow the setup instructions for use of this AI plant care assistant, Verdure AI.